// src/app/models/IArticle.ts
export interface IArticle {
    title: string;
    articleText: string;
    image: string;
    linkGerman: string;
    buttonGerman: string;
    linkSwiss: string;
    buttonSwiss: string;
    source: string;
    source_link: string;
    date: string;
    source_2: string;
    source_2_link: string;
}